Move crypto tests
